Had to remove some things, but backed them up to my desktop
***THIS IS A HUGE, IMPORTANT COMMIT***
These folders must be merged - it actually will help with circular
dependencies and other file errors. I'll go back and add the backup
stuff next, but this took me a while to do correctly so I'll take a
break and come back to that later. Long story short - this is necessary,
it may even help, just use the new namespaces for some of the classes
carefully.